@@ -9,8 +9,8 @@
     "@testing-library/user-event": "^12.1.10",
     "gh-pages": "^4.0.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
