@@ -3,6 +3,8 @@ import AppHeader from "../appHeader/AppHeader";
 import RandomChar from "../randomChar/RandomChar";
 import CharList from "../charList/CharList";
 import CharInfo from "../charInfo/CharInfo";
+import ComicsList from "../comicsList/ComicsList";
+import AppBanner from "../appBanner/AppBanner";
 import ErrorBoundary from "../errorBoundary/ErrorBoundary";
 
 import decoration from '../../resources/img/vision.png';
@@ -18,7 +20,7 @@ const App = () => {
             <div className="app">
                 <AppHeader/>
                 <main>
-                    <ErrorBoundary>
+                    {/* <ErrorBoundary>
                         <RandomChar/>
                     </ErrorBoundary>
                     <div className="char__content">
@@ -30,7 +32,9 @@ const App = () => {
                         </ErrorBoundary>
                     </div>
                     <img className="bg-decoration"
-                        src={decoration} alt="vision"/>
+                        src={decoration} alt="vision" /> */}
+                    <AppBanner/>
+                    <ComicsList/>
                 </main>
             </div>
         )
